requirements.txt
streamlit==1.29.0
groq==0.4.1
httpx==0.24.1
SpeechRecognition==3.10.0
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.1.0
PyAudio==0.2.14
python-dotenv==1.0.0
pydub==0.25.1
ffmpeg-python==0.2.0
mediapipe==0.10.9